extract displayIcon, IconConfig, proIcons to help-sidebar-icons
