+++ ui/src/app/components/help-sidebar-icons.tsx
@@ -1,22 +1,31 @@
 import * as React from 'react';
 import { CSSProperties } from 'react';
+import * as fp from 'lodash/fp';
 import { IconDefinition } from '@fortawesome/free-solid-svg-icons';
 import { faCircle } from '@fortawesome/free-solid-svg-icons/faCircle';
 import { faLock } from '@fortawesome/free-solid-svg-icons/faLock';
 import { faSyncAlt } from '@fortawesome/free-solid-svg-icons/faSyncAlt';
 import { FontAwesomeIcon } from '@fortawesome/react-fontawesome';
 
-import { RuntimeStatus } from 'generated/fetch';
+import {
+  Criteria,
+  GenomicExtractionJob,
+  RuntimeStatus,
+  TerraJobStatus,
+} from 'generated/fetch';
 
 import colors from 'app/styles/colors';
-import { reactStyles } from 'app/utils';
+import { DEFAULT, reactStyles, switchCase } from 'app/utils';
 import { ComputeSecuritySuspendedError } from 'app/utils/runtime-utils';
 import { CompoundRuntimeOpStore, RuntimeStore } from 'app/utils/stores';
 import { WorkspaceData } from 'app/utils/workspace-data';
+import { supportUrls } from 'app/utils/zendesk';
 import arrowLeft from 'assets/icons/arrow-left-regular.svg';
 import runtime from 'assets/icons/thunderstorm-solid.svg';
 import times from 'assets/icons/times-light.svg';
+import moment from 'moment/moment';
 
+import { RouteLink } from './app-router';
 import { FlexRow } from './flex';
 
 const styles = reactStyles({
@@ -35,6 +44,18 @@ const styles = reactStyles({
   },
   rotate: {
     animation: 'rotation 2s infinite linear',
+  },
+  criteriaCount: {
+    position: 'absolute',
+    height: '0.8rem',
+    width: '0.8rem',
+    top: '1rem',
+    left: '0.55rem',
+    textAlign: 'center',
+    backgroundColor: colors.danger,
+    borderRadius: '50%',
+    display: 'inline-block',
+    fontSize: '0.4rem',
   },
 });
 
@@ -55,17 +76,12 @@ export const proIcons = {
   times: times,
 };
 
-export interface DisplayRuntimeIconProps {
-  workspace: WorkspaceData;
-  runtimeStore: RuntimeStore;
-  compoundRuntimeOps: CompoundRuntimeOpStore;
-}
-
-export const displayRuntimeIcon = (
+const displayRuntimeIcon = (
   icon: IconConfig,
-  props: DisplayRuntimeIconProps
+  workspace: WorkspaceData,
+  store: RuntimeStore,
+  compoundRuntimeOps: CompoundRuntimeOpStore
 ) => {
-  const { runtimeStore: store, compoundRuntimeOps, workspace } = props;
   let status = store?.runtime?.status;
   if (
     (!status || status === RuntimeStatus.Deleted) &&
@@ -180,5 +196,231 @@ export const displayRuntimeIcon = (
         })()}
       </FlexRow>
     </FlexRow>
+  );
+};
+
+const withinPastTwentyFourHours = (epoch: number) => {
+  const completionTimeMoment = moment(epoch);
+  const twentyFourHoursAgo = moment().subtract(1, 'days');
+  return completionTimeMoment.isAfter(twentyFourHoursAgo);
+};
+
+const displayFontAwesomeIcon = (
+  icon: IconConfig,
+  criteria: Array<Selection>,
+  concept: Array<Criteria>
+) => {
+  return (
+    <React.Fragment>
+      {icon.id === 'criteria' && criteria && criteria.length > 0 && (
+        <span data-test-id='criteria-count' style={styles.criteriaCount}>
+          {criteria.length}
+        </span>
+      )}
+      {icon.id === 'concept' && concept && concept.length > 0 && (
+        <span data-test-id='concept-count' style={styles.criteriaCount}>
+          {concept.length}
+        </span>
+      )}
+      <FontAwesomeIcon
+        data-test-id={'help-sidebar-icon-' + icon.id}
+        icon={icon.faIcon}
+        style={icon.style}
+      />
+    </React.Fragment>
+  );
+};
+
+const displayExtractionIcon = (
+  icon: IconConfig,
+  genomicExtractionJobs,
+  criteria: Array<Selection>,
+  concept: Array<Criteria>
+) => {
+  const jobsByStatus = fp.groupBy('status', genomicExtractionJobs);
+  let status;
+  // If any jobs are currently active, show the 'sync' icon corresponding to their status.
+  if (jobsByStatus[TerraJobStatus.RUNNING]) {
+    status = TerraJobStatus.RUNNING;
+  } else if (jobsByStatus[TerraJobStatus.ABORTING]) {
+    status = TerraJobStatus.ABORTING;
+  } else if (
+    jobsByStatus[TerraJobStatus.SUCCEEDED] ||
+    jobsByStatus[TerraJobStatus.FAILED] ||
+    jobsByStatus[TerraJobStatus.ABORTED]
+  ) {
+    // Otherwise, show the status of the most recent completed job, if it was completed within the past 24h.
+    const completedJobs = fp.flatten([
+      jobsByStatus[TerraJobStatus.SUCCEEDED] || [],
+      jobsByStatus[TerraJobStatus.FAILED] || [],
+      jobsByStatus[TerraJobStatus.ABORTED] || [],
+    ]);
+    const mostRecentCompletedJob = fp.flow(
+      fp.filter((job: GenomicExtractionJob) =>
+        withinPastTwentyFourHours(job.completionTime)
+      ),
+      // This could be phrased as fp.sortBy('completionTime') but it confuses the compile time type checker
+      fp.sortBy((job) => job.completionTime),
+      fp.reverse,
+      fp.head
+    )(completedJobs);
+    if (mostRecentCompletedJob) {
+      status = mostRecentCompletedJob.status;
+    }
+  }
+
+  // We always want to show the DNA icon.
+  // When there are running or recently completed  jobs, we will show a small overlay icon in
+  // the bottom right of the tab showing the job status.
+  return (
+    <FlexRow
+      style={{
+        height: '100%',
+        alignItems: 'center',
+        justifyContent: 'space-around',
+      }}
+    >
+      {displayFontAwesomeIcon(icon, criteria, concept)}
+      <FlexRow
+        data-test-id='extraction-status-icon-container'
+        style={styles.statusIconContainer}
+      >
+        {switchCase(
+          status,
+          [
+            TerraJobStatus.RUNNING,
+            () => (
+              <FontAwesomeIcon
+                icon={faSyncAlt}
+                style={{
+                  ...styles.asyncOperationStatusIcon,
+                  ...styles.rotate,
+                  color: colors.asyncOperationStatus.starting,
+                }}
+              />
+            ),
+          ],
+          [
+            TerraJobStatus.ABORTING,
+            () => (
+              <FontAwesomeIcon
+                icon={faSyncAlt}
+                style={{
+                  ...styles.asyncOperationStatusIcon,
+                  ...styles.rotate,
+                  color: colors.asyncOperationStatus.stopping,
+                }}
+              />
+            ),
+          ],
+          [
+            TerraJobStatus.FAILED,
+            () => (
+              <FontAwesomeIcon
+                icon={faCircle}
+                style={{
+                  ...styles.asyncOperationStatusIcon,
+                  color: colors.asyncOperationStatus.error,
+                }}
+              />
+            ),
+          ],
+          [
+            TerraJobStatus.SUCCEEDED,
+            () => (
+              <FontAwesomeIcon
+                icon={faCircle}
+                style={{
+                  ...styles.asyncOperationStatusIcon,
+                  color: colors.asyncOperationStatus.succeeded,
+                }}
+              />
+            ),
+          ],
+          [
+            TerraJobStatus.ABORTED,
+            () => (
+              <FontAwesomeIcon
+                icon={faCircle}
+                style={{
+                  ...styles.asyncOperationStatusIcon,
+                  color: colors.asyncOperationStatus.stopped,
+                }}
+              />
+            ),
+          ]
+        )}
+      </FlexRow>
+    </FlexRow>
+  );
+};
+
+interface TempProps {
+  workspace: WorkspaceData;
+  criteria: Array<Selection>;
+  concept?: Array<Criteria>;
+  runtimeStore: RuntimeStore;
+  compoundRuntimeOps: CompoundRuntimeOpStore;
+  genomicExtractionJobs: GenomicExtractionJob[];
+}
+export const displayIcon = (icon: IconConfig, props: TempProps) => {
+  const {
+    workspace,
+    runtimeStore,
+    compoundRuntimeOps,
+    genomicExtractionJobs,
+    criteria,
+    concept,
+  } = props;
+  const terminalRoute = `/workspaces/${workspace.namespace}/${workspace.id}/terminals`;
+  return switchCase(
+    icon.id,
+    [
+      'dataDictionary',
+      () => (
+        <a href={supportUrls.dataDictionary} target='_blank'>
+          <FontAwesomeIcon
+            data-test-id={'help-sidebar-icon-' + icon.id}
+            icon={icon.faIcon}
+            style={icon.style}
+          />
+        </a>
+      ),
+    ],
+    [
+      'runtime',
+      () =>
+        displayRuntimeIcon(icon, workspace, runtimeStore, compoundRuntimeOps),
+    ],
+    [
+      'terminal',
+      () => (
+        <RouteLink path={terminalRoute}>
+          <FontAwesomeIcon
+            data-test-id={'help-sidebar-icon-' + icon.id}
+            icon={icon.faIcon}
+            style={icon.style}
+          />
+        </RouteLink>
+      ),
+    ],
+    [
+      'genomicExtractions',
+      () =>
+        displayExtractionIcon(icon, genomicExtractionJobs, criteria, concept),
+    ],
+    [
+      DEFAULT,
+      () =>
+        icon.faIcon === null ? (
+          <img
+            data-test-id={'help-sidebar-icon-' + icon.id}
+            src={proIcons[icon.id]}
+            style={icon.style}
+          />
+        ) : (
+          displayFontAwesomeIcon(icon, criteria, concept)
+        ),
+    ]
   );
 };
